refactor(frontend): extract `LoginControl`; import into forms
